#2
I made the second page, the images and text are made so the "Restaurant" owner(other side of user) can put any image and any text, as long as the image is sized to 1000x1000 pixels.

I also made it so the login will now check if your password is matching with one in the system, you can not create an account yet.